# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_convert_to_facturx_v1_convert_post": {
        "properties": {
          "metadata": {
            "description": "Invoice metadata as JSON",
            "title": "Metadata",
            "type": "string"
          },
          "pdf": {
            "contentMediaType": "application/octet-stream",
            "description": "Original PDF invoice (Bring Your Own PDF)",
            "title": "Pdf",
            "type": "string"
          }
        },
        "required": [
          "pdf",
          "metadata"
        ],
        "title": "Body_convert_to_facturx_v1_convert_post",
        "type": "object"
      },
      "Body_extract_facturx_v1_extract_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Factur-X PDF file to extract data from",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_extract_facturx_v1_extract_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 standard error response format.\nhttps://www.rfc-editor.org/rfc/rfc9457.html",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "title": "Detail"
          },
          "extensions": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Additional problem details specific to the API.",
            "title": "Extensions"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "title": "Instance"
          },
          "status": {
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference that identifies the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted REST API for technical Factur-X/ZUGFeRD and XRechnung workflows.",
    "title": "Factur-X Engine",
    "version": "1.8.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/convert": {
      "post": {
        "description": "Attach Factur-X/CII XML to an existing standard PDF invoice (BYOPDF).\n\nThe input must be a valid PDF file (max upload size is controlled by MAX_UPLOAD_SIZE_MB, default 10MB). The API generates XML from metadata and embeds it into the PDF to create a Factur-X document. Use /v1/validate to verify final PDF/A status when needed.",
        "operationId": "convert_to_facturx_v1_convert_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_to_facturx_v1_convert_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Factur-X PDF successfully generated from provided PDF"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid input"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Convert To Facturx",
        "tags": [
          "Advanced Integration"
        ]
      }
    },
    "/v1/extract": {
      "post": {
        "description": "Inspect a Factur-X PDF and return heuristic best-effort invoice data.\n\nThis endpoint is a preview/inspection workflow:\n1. Detects if the PDF contains embedded Factur-X/ZUGFeRD XML\n2. Extracts and parses the XML\n3. Returns structured invoice data (parties, totals, line items)\n\nThe response always declares `mode=preview` and\n`suitable_for_automatic_import=false`. Values may be absent, coerced, or\ntruncated by the heuristic extractor. Use `/v1/serialize` when a strict,\nversioned mapping contract is required.\n\nUse cases:\n- Human inspection and troubleshooting\n- Discovering the embedded XML and detected profile\n- Prototyping an integration before adopting a strict mapping contract",
        "operationId": "extract_facturx_v1_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_facturx_v1_extract_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid input"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Extract Facturx",
        "tags": [
          "Core Workflows"
        ]
      }
    },
...
    "/v1/serialize": {
      "post": {
        "description": "Strict, versioned JSON serialization.\n\nA 200 response means that all configured validation layers passed and the\ninvoice was mapped without XML recovery, invented values, or silently\nskipped material elements. The response still requires client-side supplier,\nduplicate, purchase-order, tax-policy, and payment checks.\n\n**Feature availability**: Builds without the strict serializer entitlement\nreturn HTTP 403. A free 30-day evaluation is linked from the project\nwebsite; no public paid checkout is currently active.",
        "operationId": "serialize_facturx_v1_serialize_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_serialize_facturx_v1_serialize_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerializationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid input"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SerializationFailureResponse"
                }
              }
            },
            "description": "Validation or strict mapping failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Serialize Facturx",
        "tags": [
          "Advanced Integration"
        ]
      }
    },
    "/v1/validate": {
      "post": {
        "description": "Validate a Factur-X PDF or XML file against EN 16931 standards.\n\nReturns a validation report with detected format, flavor, and any errors.\n\nSome legacy licensed builds also return enhanced human-readable diagnostics.\nIn every build, inspect validation completeness and the executed/skipped layers.",
        "operationId": "validate_facturx_v1_validate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_facturx_v1_validate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ValidationResult"
                    },
                    {
                      "$ref": "#/components/schemas/ProValidationResult"
                    }
                  ],
                  "title": "Response Validate Facturx V1 Validate Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid input"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Validate Facturx",
        "tags": [
          "Core Workflows"
        ]
      }
    },
    "/v1/xml": {
      "post": {
        "description": "Generate the Factur-X/CII XML content directly from JSON metadata.\n\nThis endpoint returns raw XML (Cross Industry Invoice D22B) \nwithout the PDF wrapper.",
        "operationId": "generate_facturx_xml_v1_xml_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_facturx_xml_v1_xml_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Factur-X/CII XML successfully generated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid input"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error"
          }
        },
        "summary": "Generate Facturx Xml",
        "tags": [
          "Core Workflows"
        ]
      }
    }
  },
...
}